LayoutTests/imported/w3c/web-platform-tests/html/browsers/the-windowproxy-exotic-object/document-tree-child-browsing-context-name-property-set.sub-expected.txt

Blocked access to external URL http://www.localhost:8801/common/window-name-setter.html#spices


FAIL Cross origin child window's name shadows the second candidate of a same origin iframe assert_equals: expected (undefined) undefined but got (object) object "[object Window]"
